Test all types of test account provisioning
This modifies the CI pipeline that verifies the correct functionality of this Orb. It now runs three jobs (instead of once). Each job uses a different method to provision the test accounts. At least one of the XMPP-tests (the one for XEP-0199) requires all three accounts to be present, so that's a strong indicator to determine if account provisioning works as intended.